requirement.txt
Django
rest-auth 
Djangorestframework 

